FIX : 로그인이 되어있는데 안되어있다고 표기되는 오류 수정
