+++ package.json
@@ -19,6 +19,7 @@
 		"@types/styled-components": "^5.1.26",
 		"@types/uuid": "^9.0.1",
 		"axios": "^1.3.4",
+		"bumawiki-editor": "^1.0.6",
 		"eslint": "8.35.0",
 		"eslint-config-next": "13.2.3",
 		"moment": "^2.29.4",
